api/pyproject.toml
[project]
name = "malomatia-gov-triage-api"
version = "0.2.0"
description = "Production-style API for the Malomatia Gov-Service Triage platform"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.118.0",
  "uvicorn[standard]==0.37.0",
  "sqlmodel==0.0.24",
  "psycopg[binary]==3.2.10",
  "pydantic-settings==2.10.1",
  "redis==6.4.0",
  "PyJWT==2.10.1",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
  "httpx==0.28.1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
